feat: refine team forge flow and harden agent startup input handling
